Update: add comments and clean the env related files to pass the norm
